Kerberos sign only interop issue
When kerberos is negotiated with Sign Only protection level from *nix
client against a windows server, *nix client interpretes the session flags
as supported and hence always set CONFIDENTIALITY flag irresepective of
the requested flags/protection level. This, in windows server which
interpretes the flags as required, casues the windows server to expect all
the further communication to be encrypted, which is not the case and hence
server side processing fails.
To override this behaviour, starting MIT Kerberos 1.14 and heimdal
implementaion, we need to call gss_set_cred_option with
GSS_KRB5_CRED_NO_CI_FLAGS_X. When not available, the windows server should
fail in negotiating Kerberos Sign Only.